Add Part A VAE to solution.py
