+++ solution.py
@@ -1,9 +1,46 @@
+# -*- coding: utf-8 -*-
+# ---
+# jupyter:
+#   jupytext:
+#     cell_metadata_filter: tags,title,-all
+#     custom_cell_magics: kql
+#     formats: ipynb,py:percent
+#     text_representation:
+#       extension: .py
+#       format_name: percent
+#       format_version: '1.3'
+#       jupytext_version: 1.11.2
+#   kernelspec:
+#     display_name: 07_xai
+#     language: python
+#     name: python3
+# ---
+
 # %% [markdown] tags=[]
-# # Exercise 7: XAI 
+# # Exercise 7: Representation learning and XAI
 #
-# The goal of this exercise is to learn how to probe what a pre-trained classifier has learned about the data it was trained on.
+# Representation learning has been gaining significant attention over the past years,
+# driven by the rise of new architectures and the growing need in biology to extract meaningful structured information
+# from increasingly high-dimensional and complex data. Yet as these models become more capable, understanding what they are actually learning becomes
+# just as important as their predictive performance, and that is precisely the role of explainable AI.
 #
-# We will be working with a simple example which is a fun derivation on the MNIST dataset that you will have seen in previous exercises in this course.
+# Goal of the exercises
+#
+# The goal of this exercise is to first build an understanding of what representation learning is,
+# what we mean by a "representation" in that context, and what makes a representation good or useful.  
+# From there, we will explore what architectures can be used to obtain these representations, and how we can evaluate
+# the quality of the obtained representations.  The second half of the exercise shifts focus to Explainable AI (XAI),
+# where the goal is to learn how to probe what a pre-trained classifier has learned about the data it was trained on.
+#
+# In part A, we will be building models for representation learning 
+#
+# We will:
+# 1. Build and train a variational autoencoder (VAE)
+# 2. Visualize and compare the latent space learned under different model parameters
+# 3. Evaluate representation quality 
+# 4. Explore the generative properties of VAEs. 
+#
+# In part B, we will be working with a simple example which is a fun derivation on the MNIST dataset that you will have seen in previous exercises in this course.
 # Unlike regular MNIST, our dataset is classified not by number, but by color!
 #
 # We will:
@@ -13,19 +50,1604 @@
 # 4. Create attributions from the counterfactual, and learn the differences between the classes.
 #
 # If time permits, we will try to apply this all over again as a bonus exercise to a much more complex and more biologically relevant problem.
+#
 # ### Acknowledgments
 #
-# This notebook was written by Diane Adjavon, with input from Alex Hillsley, Ed Hirata, Larissa Heinrich, Morgan Schwartz, Anna Foix, and Ben Salmon. 
-# It was inspired by a previous version written by Jan Funke and modified by Tri Nguyen, using code from Nils Eckstein.
+# This notebook was written by Diane Adjavon, Maria Theiss and Anna Foix-Romero with input from
+# Alex Hillsley, Ed Hirata, Larissa Heinrich, Morgan Schwartz, Anna Foix-Romero, Ben Salmon, Albert Dominguez, Talley Lambert and Eva de la Serna.
+# Part B was inspired by a previous version written by Jan Funke and modified by Tri Nguyen, using code from Nils Eckstein.
+# Part A has been inspired by multiple discussions between Virginie Uhlhmann, Alex Krull, Martin Weigert,
+# Albert Dominguez, Ed Hirata and Anna Foix-Romero.  
 #
+# ### AI Statement
+#
+# Portions of this notebook were developed with the assistance of **Claude 4.6** (Anthropic),
+# accessed via the **Harvard AI Sandbox**. Prompts and responses are not used to train external models and data does not leave the
+# university's controlled infrastructure.
+#
+# Specifically, Claude was used for:
+# - **Inspiration and structure** of explanatory markdown text 
+# - **Debugging and refinement** of plotting and training code
+# - **Drafting** of reusable helper functions
+#
+# All AI-assisted content was reviewed, edited, and validated by the notebook authors.
+#
+# ***
 # %% [markdown]
 # <div class="alert alert-danger">
 # Set your python kernel to <code>07_xai</code>
 # </div>
-
+#
+# %% [markdown]
+# # PART A: Representation learning
+# ## Part A.0: Conceptual introduction 
+# ### What is a representation and why is that useful?
+# A representation is a mapping from raw data to a structured, typically lower-dimensional space, called latent space, that ideally captures meaningful features of the data.  
+#
+# Representations  
+# - Can compress the data 
+# - Ideally learn meaningful features and discard noise and redundancy 
+# - Enable classification, clustering, and data generation 
+# - Aid interpretability 
+#
+# ### What is a good representation?
+# In many cases, good representations are... 
+#
+# **Compact**  
+# Later, we will see example images of MNIST – 28 x 28 images of hand-written digits on dark background.  
+# We can view each image as existing in an 28 x 28 = 784 - dimensional space. However, most of the image carries irrelevant information. 
+# For instance, most pixels belong to the dark background that does not contain infromation about the imaged digit.  
+# By keeping a representation compact (but not too compact!), we force the model to discard irrelevant information. 
+#
+# **Smooth and continuous**  
+# Small changes in the input should lead to small changes in the representation. There should be no "gaps" in the representation.  
+# This property makes it possible to generate new data from the representation, as each location in the latent space contains meaningful information.  
+#
+# **Structured**  
+# Similar inputs should be mapped to close-by locations in latent space, whereas dissimilar inputs should be distant to each other in latent space. 
+# This means that clusters present in the input, should also appear in latent space. 
+#
+# **Disentangled**  
+# Ideally, each dimension of the latent space should capture a different feature of the input-image. For example, one dimension could capture the tilt (left-right) of handwritten digits, 
+# another the identity of the digits, yet another the boldness of the text.  
+#
+# ### Unsupervised learning
+# The Variational Autoencoder (VAE) trained in part A is an example of representation learning and unsupervised learning. We do not provide labels – i.e. the model trains on images of hand-written digits, but does not know the true identity (labels 0 - 9) of each image. 
+# This means the model is learning structural information that is intrinsic to the data. It does so by fulfilling two training objectives: 
+# - Reconstruction: Reconstruction-loss incentivises the model to generate a reconstructed version of an input-image from the latent space. This means that the latent space needs to ideally carry information to allow for a close reconstruction.  
+# - Constraints on the latent space: We can impose structural constraints on the latent space, ensuring previously discussed smoothness and continuity. 
+#
+# Unsupervised learning is valuable in applications where labels are costly to generate or entirely unknown. 
+#
+# ---
 # %% [markdown]
 #
-# # Part 1: Setup
+# ## Part A.1: General set-up
+# In this part of the notebook, we will load the same dataset as in the previous exercise. 
+#
+# ### Background - the MNIST dataset
+# MNIST is a machine learning benchmark dataset:
+# * **70,000** grayscale images of handwritten digits **0 - 9**.  
+# * Of which are **60,000** training images and **10,000** testing images. 
+# *  Each image has a resolution of **28x28** pixels.  
+#
+# It is a great dataset to introduce representation learning because it is simple enough to train quickly,
+# but still structured enough that we can visually inspect and intuitively evaluate the quality of the learned representations
+# and reconstructions.
+#
+# Documentation for this pytorch dataset is available at https://pytorch.org/vision/main/generated/torchvision.datasets.MNIST.html
+#
+# Let's get started and load our dataset, transforming the images into torch tensors and normalising them.
+
+# %% [markdown]
+# ### Load MNIST
+
+# %%
+import torchvision
+
+transform = torchvision.transforms.Compose([torchvision.transforms.ToTensor()])
+train_mnist = torchvision.datasets.MNIST("./mnist", train=True, download=False, transform=transform)
+test_mnist = torchvision.datasets.MNIST("./mnist", train=False, download=False, transform=transform)
+
+# %% [markdown]
+# <div class="alert alert-info">
+#     <b>Note:</b> set the <code>download</code> argument of <code>torchvision.datasets.MNIST</code> to <code>True</code> or <code>False</code> as required when re-running the notebook. <br>
+#     When <code>./mnist</code> does not yet exist (on first run), make sure the first call to <code>torchvision.datasets.MNIST</code> has <code>download</code> set to <code>True</code>.
+# </div>
+
+# %% [markdown]
+# #### Inspect the train data
+# Let's take a look at a few loaded samples:
+
+# %%
+import matplotlib.pyplot as plt
+
+# Show some examples
+fig, axs = plt.subplots(4, 4, figsize=(8, 8))
+
+# Load the first 16 images and labels
+xs = [train_mnist[i][0] for i in range(16)] # images
+ys = [train_mnist[i][1] for i in range(16)] # labels
+
+for i, ax in enumerate(axs.flatten()):
+    x = xs[i]
+    y = ys[i]
+    x = x.permute((1, 2, 0))  # make channels last
+    im = ax.imshow(x,  cmap = "gray")
+    ax.set_title(f"Class {y}")
+    ax.axis("off")
+
+fig.colorbar(im, ax=axs, orientation='vertical', label="gray value", shrink = 0.9)
+# %% [markdown]
+# #### Dataloaders
+# Now, from the loaded datasets (both the train and test splits), we derive the dataloaders. We use dataloaders as they provide additional load-time features.  
+# Specifically, a dataloader enables **iterating** over the dataset in batches. It provides **shuffling** if desired.  
+# Here, we set the `batch_size` for both the train and test loader, and set `shuffle` for training only.
+
+# %%
+from torch.utils.data import DataLoader
+
+batch_size = 8
+train_loader = DataLoader(train_mnist, batch_size=batch_size, shuffle=True)
+test_loader = DataLoader(test_mnist, batch_size=batch_size, shuffle=False)
+
+# %% [markdown]
+# The dataset iterator and the dataloader iterator differ in shape: 
+
+# %%
+# Dataset iterator
+smpl, lbl = next(iter(train_mnist))
+print(f"dataset element shape: {smpl.shape} (class: {lbl})")
+
+# Dataloader iterator
+smpl, lbl = next(iter(train_loader))
+print(f"dataloader element shape: {smpl.shape} (class: {lbl})")
+
+# %% [markdown]
+# Dataloader elements come in **8** at a time, which is the `batch_size` we set above. Hence, the first tensor dimension is the "batch" dimension and has size **8**.   
+# The labels are in a tensor of size **8** as opposed to a single value like in the dataset case.  
+# Note that, both in the dataset and in the dataloader, the data is not presented as a 2d `28x28` image, but rather as `1x28x28` 3d piece of data.  
+# This is useful when using multichannel data, but in our case, this extra dimension is superfluous. We will therefore drop the channel dimension for training. 
+#
+# In Summary: 
+# | | Dataset | DataLoader |
+# |---|---|---|
+# | Image shape | `(1, 28, 28)` | `(8, 1, 28, 28)` |
+# | Dimensions  | Channel, Height (Y), Width (X) | Batch, Channel, Height (Y), Width (X) |
+# | Label | scalar `int` | `tensor` of size `batch_size` (8) |
+# | Batch dimension | ❌ | ✅ (size = `batch_size`) |
+
+# %% [markdown]
+# ## Part A.2: Variational Autoencoders (VAE)
+# A Variational Autoencoder (VAE) is a machine learning architecture capable of learning a compressed representation of data by pushing it through a low-dimensional "bottleneck" and then expanding it back into its original size.  
+# The model is forced to rebuild with limited information, and must therefore learn to capture only the most important features, performing non-linear dimensionality reduction.  
+# In our case, we convert `28 * 28 = 784` pixel images into a **few** core features via the encoder part of the model. The decoder part then turns these few features back into `28 * 28` pixel images.
+#
+# ### Part A.2.1: The architecture of our VAE
+# #### A shared backbone for encoder and decoder
+# We choose a **Multilayer Perceptron (MLP)** as the architecture to back both the **encoder** and the **decoder**.  
+# MLPs consist of linear transformations (weights and biases) followed by non-linear activation functions (ReLU). 
+
+# %%
+import torch.nn as nn
+
+class MLP(nn.Module):
+    def __init__(self, input_dim, output_dim
+                , hidden_dims=[], activation=nn.ReLU(), final_activation=None ):
+        
+        super().__init__()
+        dims = [input_dim] + hidden_dims
+        layers = []
+
+        for i in range(len(dims) - 1):
+            layers.append(nn.Linear(dims[i], dims[i + 1]))
+            if activation is not None: layers.append(activation)
+
+        layers.append(nn.Linear(dims[-1], output_dim))
+
+        if final_activation is not None:
+            layers.append(final_activation)
+
+        self.net = nn.Sequential(*layers)
+
+
+    def forward(self, x):
+        return self.net(x)
+# %% [markdown]
+# #### Building the `VariationalAutoEncoder` class
+#
+# Here we provide a `VariationalAutoEncoder` class that uses the previously defined **MLP** class for the **encoder** and **decoder**.  
+#
+# **The encode function**: 
+# * Reshapes the input by dropping the unused channel dimension. 
+# * Reduces width and height (28 x 28) to a single dimension (28 * 28 = 784). 
+# * Outputs `mu` and `logvar`; single-dimensional tensors that represent the center and logvariance for a given input in the latent space. 
+#
+# **The reparametrization trick**: 
+# * Allows to sample the latent variable `z` as it if came from a Normal distribution with `mu` and $std = e^{logvar/2}$  
+# * However, sampling `z` from `mu` and `std` does not allow for backpropagation
+# * To allow for backpropagation, the model samples $\epsilon$ from a Normal distritubion with mean 0 and standard deviation 1
+# * And then samples z with the help of $\epsilon$: $z = mu + \epsilon * e^{logvar/2}$
+#
+# **The decode function**: 
+# * Takes in a latent vector `z` and uses an MLP to reconstruct the original sample, reshaping as appropriate.
+# %% [markdown]
+# <div class="alert alert-block alert-info"><h2>Task – Fill in the gaps</h2>
+#
+# There are gaps marked as `...`.  
+# Fill them in:
+#
+# **Missing in `def __init__`**
+#
+# The encoder and decoder instance of the MLP are missing. Replace `...` with:     
+# `self.encoder`  
+# `self.decoder`. 
+#
+# How can you tell which is which?
+#
+#
+# **Missing in `def reparameterize`**  
+#
+# `epsilon` (missing twice)  
+# `std` 
+#
+# Tip:  
+# $std = e^{logvar/2}$   
+# $z = mu + \epsilon * std$
+#
+#
+# **Missing in `def forward`**  
+# `mu`   
+# `logvar`   
+# `z`  
+# `self.decode`  
+# `self.encode`  
+# `self.reparameterize`  
+# `xx` (this is the reconstructed image) 
+#
+# </div>
+#
+#
+
+# %%
+import torch
+
+class VariationalAutoEncoder(nn.Module):
+    def __init__( self, w, h, latent_dim
+                , enc_hidden_dims=[256, 128, 64], enc_activation=nn.ReLU()
+                , dec_hidden_dims=[64, 128, 256], dec_activation=nn.ReLU()
+                ):
+        super().__init__()
+        self.w = w
+        self.h = h
+        data_dim = w * h
+        
+        # TODO 
+        ... = MLP( data_dim, latent_dim * 2 # latent_dim * 2, because it will be split into mu and logvar
+                          , hidden_dims=enc_hidden_dims
+                          , activation=enc_activation)
+        
+        # TODO  
+        ... = MLP( latent_dim, data_dim
+                          , hidden_dims=dec_hidden_dims
+                          , activation=dec_activation
+                          , final_activation=nn.Sigmoid())
+
+
+        
+
+
+
+    def encode(self, x):
+        b  = x.shape[0] # batchsize
+        x_flat = x.reshape(b, -1) # (8, 1, 28, 28) → (8, 784)
+        out = self.encoder(x_flat)
+        mu, logvar = torch.chunk(out, 2, dim = 1)
+        return mu, logvar
+    
+    @staticmethod
+    def reparameterize(mu, logvar):
+        ... = torch.exp(logvar / 2) # TODO 
+        ... = torch.randn_like(std) # TODO 
+        return ... * std + mu # TODO 
+
+    def decode(self, z):
+        out = self.decoder(z) 
+        xx = out.reshape(-1, 1, self.w, self.h) # (8, 784) → (8, 1, 28, 28)
+        return xx
+    
+    def forward(self, x):
+        ..., ... = ...(x) # TODO
+        ... = ...(mu, logvar)     # TODO   
+        ... = ...(z)           # TODO      
+        return xx, z, mu, logvar
+
+# %% tags=["solution"]
+import torch
+
+class VariationalAutoEncoder(nn.Module):
+    def __init__( self, w, h, latent_dim
+                , enc_hidden_dims=[256, 128, 64], enc_activation=nn.ReLU()
+                , dec_hidden_dims=[64, 128, 256], dec_activation=nn.ReLU()
+                ):
+        super().__init__()
+        self.w = w
+        self.h = h
+        data_dim = w * h
+        
+        self.encoder = MLP( data_dim, latent_dim * 2 # latent_dim * 2, because it will be split into mu and logvar
+                          , hidden_dims=enc_hidden_dims
+                          , activation=enc_activation)
+        self.decoder = MLP( latent_dim, data_dim
+                          , hidden_dims=dec_hidden_dims
+                          , activation=dec_activation
+                          , final_activation=nn.Sigmoid())
+        
+
+    def encode(self, x):
+        b  = x.shape[0] # batchsize
+        x_flat = x.reshape(b, -1) # (8, 1, 28, 28) → (8, 784)
+        out = self.encoder(x_flat)
+        mu, logvar = torch.chunk(out, 2, dim = 1)
+        return mu, logvar
+    
+    @staticmethod
+    def reparameterize(mu, logvar):
+        std = torch.exp(logvar / 2)
+        epsilon = torch.randn_like(std)
+        return epsilon * std + mu
+
+    def decode(self, z):
+        out = self.decoder(z) 
+        xx = out.reshape(-1, 1, self.w, self.h) # (8, 784) → (8, 1, 28, 28)
+        return xx
+    
+    def forward(self, x):
+        mu, logvar = self.encode(x)
+        z = self.reparameterize(mu, logvar)
+        xx = self.decode(z)
+        return xx, z, mu, logvar
+# %% [markdown]
+# Run the following tests to confirm: 
+
+# %%
+import inspect
+import re
+
+def normalize(s):
+    return re.sub(r'\s+', '', s)
+
+def test_vae(w=28, h=28, latent_dim=16, batch_size=8):
+    vae = VariationalAutoEncoder(w, h, latent_dim)
+    data_dim = w * h
+    x = torch.randn(batch_size, 1, w, h)
+    
+    #  __init__: self.encoder
+    assert hasattr(vae, 'encoder'), \
+        "❌ 'self.encoder' missing or misspelled — correct it in __init__"
+    try:
+        enc_out = vae.encoder(torch.randn(batch_size, data_dim))
+    except RuntimeError:
+        raise AssertionError(
+            f"❌ self.encoder crashed on input shape ({batch_size}, {data_dim}) "
+            f"— did you swap self.encoder and self.decoder?"
+        )
+    assert enc_out.shape == (batch_size, latent_dim * 2), \
+        f"❌ encoder output {enc_out.shape} ≠ ({batch_size}, {latent_dim*2}) — did you swap encoder/decoder?"
+    print("✅ self.encoder: correct")
+
+    #  __init__: self.decoder
+    assert hasattr(vae, 'decoder'), \
+        "❌ 'self.decoder' missing or misspelled — correct it in __init__"
+    try:
+        dec_out = vae.decoder(torch.randn(batch_size, latent_dim))
+    except RuntimeError:
+        raise AssertionError(
+            f"❌ self.decoder crashed on input shape ({batch_size}, {latent_dim}) "
+            f"— did you swap self.encoder and self.decoder?"
+        )
+    assert dec_out.shape == (batch_size, data_dim), \
+        f"❌ decoder output {dec_out.shape} ≠ ({batch_size}, {data_dim}) — did you swap encoder/decoder?"
+    print("✅ self.decoder: correct")
+
+    #  reparameterize: check variable names 
+    norm_reparam = normalize(inspect.getsource(VariationalAutoEncoder.reparameterize))
+
+    assert normalize('std = torch.exp(logvar / 2)') in norm_reparam, \
+        "❌ reparameterize: first blank wrong"
+    print("✅ reparameterize: 'std' correct")
+
+    assert normalize('epsilon = torch.randn_like(std)') in norm_reparam, \
+        "❌ reparameterize: second blank wrong"
+    print("✅ reparameterize: 'epsilon' correct")
+
+    assert  normalize('return epsilon * std + mu') in norm_reparam or \
+            normalize('return std * epsilon + mu') in norm_reparam or \
+            normalize('return mu + epsilon * std') in norm_reparam or \
+            normalize('return mu + std * epsilon') in norm_reparam, \
+            "❌ reparameterize: third blank wrong"
+    print("✅ reparameterize: return statement correct")
+
+    # forward: check variable names
+    xx, z, mu, logvar = vae(x)
+    assert xx.shape == (batch_size, 1, w, h), "❌ forward failed at runtime"
+    norm_forward = normalize(inspect.getsource(VariationalAutoEncoder.forward))
+
+    assert normalize('mu, logvar = self.encode(x)')         in norm_forward, \
+        "❌ forward: check line 1"
+    assert normalize('z = self.reparameterize(mu, logvar)') in norm_forward, \
+        "❌ forward: check line 2"
+    assert normalize('xx = self.decode(z)')                 in norm_forward, \
+        "❌ forward: check line 3"
+    print("✅ forward: all blanks correct")
+
+    print("\n All tests passed!")
+
+test_vae()
+
+# %% [markdown]
+# <div class="alert alert-block alert-success"><h2>Checkpoint</h2>
+#
+#
+# </div>
+
+# %% [markdown]
+# ### Part A.2.2: The loss functions 
+# Training a VAE balances two competing objectives: 
+#
+# #### Reconstruction  
+# The reconstruction loss measures how well the decoder reconstructs an input image from the latent space.  
+# The **Binary Cross Entropy** reconstruction loss (**BCE loss**) is appropriate here, as we scale input images to gray values of [0, 1].  
+# The decoder's final activation is a Sigmoid function, mapping to the same scale of [0, 1]. 
+#
+#
+# #### Latent space regularization  
+# **Kullback-Leibler divergence** loss (**KL loss**) measures how much a learned distribution of images in the latent space diverges from a standard normal distribution, i.e. a Normal distribution with mean 0 and std 1.  
+# KL-loss penalizes the latent space distribution for being different from a standard normal. 
+# %%
+# The reconstruction loss
+rec_loss = nn.BCELoss(reduction="sum")
+
+
+# The KL loss
+def kl_loss(mu, logvar):
+    # sum over latent dimensions, mean over batch
+    return torch.mean(-0.5 * torch.sum(1 + logvar - mu**2 - logvar.exp(), dim=1)) 
+
+
+# %% [markdown]
+# #### Combined loss with beta weighting  
+# We combine the reconstruction loss and KL loss into an overall loss. The overall loss is what we backpropagate on.  
+# Parameter `beta` adds a weighting to the KL loss.  
+# `beta = 0` means only the reconstruction-loss influences training. 
+
+# %% [markdown]
+# <div class="alert alert-block alert-info"><h2>Task – Overall loss</h2>
+# Write code that weights the kl loss by beta and adds it to the reconstruction loss  
+# </div>
+
+# %%
+def loss(rec, kl, beta):
+    return ... 
+
+
+# %% tags=["solution"]
+def loss(rec, kl, beta):
+    return rec + beta * kl
+
+
+# %% [markdown]
+# <div class="alert alert-block alert-success"><h2>Checkpoint</h2>
+#
+#
+# </div>
+
+# %% [markdown]
+# ### Part A.2.3: Training infrastructure
+# Now we get to create and train our model on the MNIST dataset.
+#
+# #### A.2.3.1 Set the device 
+# As our model and dataset is small, CPUs are likely to outperform GPUs.  
+# The overhead of transferring the data to GPU might make the model slower than running it on CPU. 
+# %%
+device = torch.device("cpu") 
+
+# %% [markdown]
+# #### Part A.2.3.2: Model instance and optimizer
+# **Model instance**  
+# We first create an instance of our `VariationalAutoEncoder` class. On construction, it needs to know the size of the data it will receive and the desired latent space size.  
+# We grab a sample from the dataset to derive the appropriate input size.  
+# We set the latent dimensions to **2**. A word of warning on this: 
+
+# %% [markdown]
+# <div class="alert alert-info">
+# <b> Attention: </b><code>latent_dim = 2 </code> is likely underparametrizing the latent space. We will lose a lot of relevant information when compressing a 28 x 28 image down to 2 dimensions. <br>
+# However, we will roll with this for now as it helps us build an intution for the latent space.  <br>
+# Later, we will retrain with a higher-dimensional latent space.  
+# </div>
+
+# %%
+data_sample, _ = next(iter(train_mnist))
+_, w, h = data_sample.shape
+model = VariationalAutoEncoder(w, h, latent_dim=2).to(device)
+
+# %% [markdown]
+# **Optimizer**  
+# We then create an optimizer for the model's parameters. During training, gradients will be computed from the backpropagation pass.  
+# The optimizer will read those gradients and update the model's parameters.
+
+# %%
+from torch.optim import Adam
+optimizer = Adam(model.parameters(), lr=0.0001)
+
+# %% [markdown]
+# #### Part A.2.3.3: The training "loop"
+# To train a model, the general idea is to iterate through the dataset, passing each element through the model to produce a reconstruction and embed it into the latent space.  
+# We observe how close to the original data and the reconstruction is using the reconstruction loss function, and use that observation to inform the model optimisation.  
+# We penalize the latent space for not following a standard normal distribution using the KL-loss.  
+# Performing these steps going once through all the training data is what is referred to as a training **epoch**. 
+# We then loop this process over for a desired arbitrary number of training epochs.  
+#
+# Below are three functions:  
+# `train_epoch`: capture training for a single epoch and which returns the average epoch loss.  
+# `train_epochs`: Calls `train_epoch` for the desired number of epochs and returns the losses per epoch.  
+# `plot_losses_live`: Plots losses during training and live-updates every few epochs.   
+# %%
+from tqdm.auto import tqdm
+from itertools import islice
+from IPython.display import clear_output
+import matplotlib.pyplot as plt
+
+
+def train_epoch(model, loader, optimizer, loss, beta):
+    model.train()
+    running_rec_loss = 0.0
+    running_kl_loss = 0.0
+    running_loss = 0.0
+
+    for x, _ in loader:
+        x = x.to(device)
+        xx, _, mu, logvar = model(x)
+        rec_l = rec_loss(xx, x)
+        kl_l = kl_loss(mu, logvar)
+        l = loss(rec_l, kl_l, beta = beta)
+
+        optimizer.zero_grad()
+        l.backward()
+        optimizer.step()
+
+        # Stats
+        running_rec_loss += rec_l.item()
+        running_kl_loss += kl_l.item()
+        running_loss += l.item()
+
+    avg_rec_loss = running_rec_loss / len(loader)
+    avg_kl_loss = running_kl_loss / len(loader)
+    avg_loss = running_loss / len(loader)
+
+    return avg_rec_loss, avg_kl_loss, avg_loss
+
+
+
+def plot_losses_live(epoch_losses, epoch_rec_losses, epoch_kl_losses):
+    _, axes = plt.subplots(1, 3, figsize=(15, 3))
+
+    for ax, values, title in zip(
+        axes,
+        [epoch_losses, epoch_rec_losses, epoch_kl_losses],
+        ["Total Loss", "Reconstruction Loss", "KL Loss"]
+    ):
+        ax.plot(values, c = "k")
+        ax.set_title(f"{title}: {values[-1]:.4f}")
+        ax.set_xlabel("Epoch")
+        ax.grid(True, linestyle="--", alpha=0.6)
+
+    plt.tight_layout()
+    plt.show()
+
+
+def train_epochs(n, model, loader, optimizer, loss, beta, plot_every=10):
+    epoch_rec_losses = []
+    epoch_kl_losses = []
+    epoch_losses = []
+
+    for epoch in range(n):
+        fresh_loader_iter = iter(loader)
+        sliced_loader = tqdm(islice(fresh_loader_iter, 100), total=100, disable=True)
+        avg_rec_loss, avg_kl_loss, avg_loss = train_epoch(
+            model, sliced_loader, optimizer, loss, beta=beta
+        )
+        epoch_rec_losses.append(avg_rec_loss)
+        epoch_kl_losses.append(avg_kl_loss)
+        epoch_losses.append(avg_loss)
+
+        if (epoch + 1) % plot_every == 0:
+            clear_output(wait=True)
+            print(f"Epoch {epoch+1}/{n} | loss={avg_loss:.4f} | rec={avg_rec_loss:.4f} | kl={avg_kl_loss:.4f}")
+            plot_losses_live(epoch_losses, epoch_rec_losses, epoch_kl_losses)
+
+    return {"loss": epoch_losses, "rec": epoch_rec_losses, "kl": epoch_kl_losses}
+
+
+# %% [markdown]
+# ### Part A.2.4: Train a model for 100 epochs
+
+# %% [markdown]
+# We can now train `model`.  
+# In this part, we want to understand the model rather than analyze its results.  
+# Therefore, we sacrifice quality for time by only training 100 epochs.  
+
+# %%
+epochs = 100
+train_epochs(epochs, model, train_loader, optimizer, loss, beta = 1);
+
+# %% [markdown]
+# ### Part A.2.5: Inspect the trained model
+# #### Get a test image
+# Let's inspect what each part of the model does.  
+# We will probe the model with **one** image.  
+# Let's get it from the `test_loader`:
+
+# %%
+x, _ = test_loader.dataset[5] # reminder: test_loader is the dataloader for the test data 
+
+print(f"Image shape : {x.shape} --- C, Y, X") # Just one image! 
+
+plt.figure(figsize=(1,1))
+plt.imshow(x.squeeze(), cmap="Grays");
+
+# %% [markdown]
+# #### The encoder
+# Next, we pass the loaded image to the encoder, which returns `mu` and `logvar`.  
+# We previously set the latent space to two dimensions. Therefore, each `mu` and `logvar` have two entries. 
+#
+# Let's have a look at `mu` and `logvar` 
+
+# %%
+model.eval()
+with torch.no_grad():
+    mu, logvar = model.encode(x.unsqueeze(0).to(device))
+
+
+print(f"mu    : {mu}. Dimensions: {len(mu[0])}")
+print(f"logvar: {logvar}. Dimensions: {len(logvar[0])}")
+
+# %% [markdown]
+# <div class="alert alert-block alert-info"><h3>Task: Compare size of an inpute image to the latent space</h3>
+# The function <code>.nbytes</code> allows you to see the size of an array or tensor in bytes.  <br>
+# Print the size of <code>x</code>.  <br>
+# Compute the size of <code>mu</code> and the size of <code>logvar</code>. Print their sum.  <br>
+# Which is bigger: the input image, or its latent space embedding? 
+
+# %%
+#Example: 
+example_tensor = torch.zeros(1)
+print(example_tensor.nbytes)
+
+# TODO: 
+...
+
+# %% tags=["solution"]
+example_tensor = torch.zeros(1)
+print(example_tensor.nbytes)
+
+print(f"Input image: {x.nbytes} bytes")
+print(f"mu + logvar: {mu.nbytes + logvar.nbytes} bytes")
+
+# %% [markdown]
+# #### Sample
+# Values `z` are sampled from a Normal distribution with mean `mu` and standard deviation $\sigma = e^{\,\text{logvar}/2}$  
+# The `reparametrize` function allows for this sampling without blocking backpropagation.  
+# Note how `z` is different for each draw. Here, we draw twice and call the result `z_1` and `z_2`. 
+
+# %%
+with torch.no_grad():
+    z_1 = model.reparameterize(mu, logvar)
+    z_2 = model.reparameterize(mu, logvar) # same mu, same logvar
+
+print(f"z_1: {z_1}, Dimensions: {len(z_1[0])}")
+print(f"z_2: {z_2}, Dimensions: {len(z_2[0])}")
+print("Are the two samples the same?", z_1 == z_2 )
+
+
+# %% [markdown]
+# #### The decoder
+# Next, we can decode `z_1` and `z_2`.  
+# The decoder reconstructs image `rec_1` from `z_1`, and image `rec_2` from `z_2`
+
+# %%
+
+with torch.no_grad():
+    rec_1 = model.decode(z_1).cpu().numpy().squeeze()
+    rec_2 = model.decode(z_2).cpu().numpy().squeeze()
+
+# %% [markdown]
+# Are the two reconstructed images the same?
+
+# %%
+fig, ax = plt.subplots(1, 4)
+
+im0 = ax[0].imshow(x.squeeze(), cmap= "Grays")
+ax[0].set_title("original")
+plt.colorbar(im0, shrink = 0.2)
+
+im0 = ax[1].imshow(rec_1, cmap= "Grays")
+ax[1].set_title("rec 1")
+plt.colorbar(im0, shrink = 0.2)
+
+im1 = ax[2].imshow(rec_2, cmap= "Grays")
+ax[2].set_title("rec 2")
+plt.colorbar(im1, shrink = 0.2)
+
+im2 = ax[3].imshow(rec_1 - rec_2, cmap= "Grays")
+ax[3].set_title("rec 1 - rec 2")
+plt.colorbar(im2, shrink = 0.2)
+
+plt.tight_layout()
+
+
+# %% [markdown]
+# Clearly not! VAEs are *probabilistic* models. 
+
+# %% [markdown]
+# #### Display multiple reconstructions
+# Now let's look at what reconstructions look like for multiple images.  
+# Below are a couple simple visualisation function to display original and reconstructed images together, and to query the model for a batch of reconstructions and display them using the first function.
+
+# %%
+def show_recon(og, recon):
+    """
+    og and recon: Tensors of shape (B, 1, 28, 28) or (B, 784)
+    """
+    b, c, h, w = og.shape
+    og = og.view(-1, h, w).detach().cpu()
+    recon = recon.view(-1, h, w).detach().cpu()
+
+    fig, axes = plt.subplots(nrows=2, ncols=b, figsize=(b * 1.5, 3))
+
+    for i in range(b):
+        # Top row: Original
+        axes[0, i].imshow(og[i], cmap='gray')
+        axes[0, i].axis('off')
+        if i == 0: axes[0, i].set_title("Original")
+
+        # Bottom row: Reconstruction
+        axes[1, i].imshow(recon[i], cmap='gray')
+        axes[1, i].axis('off')
+        if i == 0: axes[1, i].set_title("Recon")
+
+    plt.show()
+
+#import torch
+def view_test_sample(model, loader):
+    model.eval()
+    with torch.no_grad():
+        images, _ = next(iter(loader))
+        images = images.to(device)
+        recon, _, _, _ = model(images)
+    show_recon(images, recon)
+
+
+# %% [markdown]
+# Let's call this on our model as it currently stands.
+
+# %%
+view_test_sample(model, test_loader)
+
+# %% [markdown]
+# Not great, not terrible... After the checkpoint, we will instantiate new models and train them for longer.  
+
+# %% [markdown]
+# <div class="alert alert-block alert-success"><h2>Checkpoint</h2>
+#
+#
+#
+#
+
+# %% [markdown]
+# ### Part A.2.6: Train two models for 1000 epochs 
+
+# %% [markdown]
+# #### A.2.6.1: Train a model without regularized latent sapce
+
+# %% [markdown]
+# <div class="alert alert-block alert-info"><h2>Task</h2>
+#
+# Let's train our first "serious" model. 
+# * Instantiate a new variational autoencoder model and name it `model0`
+# * Keep `latent_dim = 2`. This is not ideal, but helps us better understand the latent space. 
+# * Instantiate a new optimizer 
+# * Pass `beta = 0`
+# * Train your new model for `epochs = 1000`
+#
+#
+#
+# </div>
+#
+
+# %%
+
+model0 = VariationalAutoEncoder(...).to(device)  # TODO 
+optimizer = Adam(model0.parameters(), lr=0.0001)         # fresh optimizer
+... 
+... 
+losses0 = train_epochs(epochs, model0, train_loader, optimizer, loss, beta = beta)
+
+# %% tags=["solution"]
+
+model0 = VariationalAutoEncoder(w, h, latent_dim = 2).to(device)  # fresh weights
+optimizer = Adam(model0.parameters(), lr=0.0001)         # fresh optimizer
+
+epochs = 1000
+beta = 0
+losses0 = train_epochs(epochs, model0, train_loader, optimizer, loss, beta = beta)
+
+# %% [markdown]
+# Let's have a look at the results. Reconstructed images should now look better. 
+
+# %%
+view_test_sample(model0, test_loader)
+
+# %% [markdown]
+# <div class="alert alert-block alert-warning"><h4> Questions </h4>
+#
+# * Which loss is decreasing more? Why? 
+#
+# </div>
+
+# %% [markdown]
+# <div class="alert alert-block alert-success"><h2>Checkpoint</h2>
+#
+#
+# </div>
+
+# %% [markdown]
+# #### A.2.6.2: Train a model with regularized latent sapce 
+# We are now training another model called `model1`
+
+# %% [markdown]
+# <div class="alert alert-block alert-info"><h2>Task</h2>
+#
+# Change one variable in the code below to train a new model with better-behaved KL loss.  
+# In this case, the KL loss doesn't have to decrease – we just want it to increase less.  
+#
+# Tips: 
+# * Have a look at the overall loss function definitions 
+# * Look at the order of magnitude of the reconstruction loss and KL loss, for instance at epoch 1000, to decide on a value
+# * You can train for fewer epochs if you want to try multiple values. Train for 1000 epochs once you decided 
+# </div>
+
+# %%
+model1 = VariationalAutoEncoder(w, h, latent_dim=2).to(device)
+optimizer = Adam(model1.parameters(), lr=0.0001)         # fresh optimizer
+epochs = 1000
+beta = 0
+losses1 = train_epochs(epochs, model1, train_loader, optimizer, loss, beta = beta)
+
+
+
+
+# %% tags=["solution"]
+# beta 1
+model1 = VariationalAutoEncoder(w, h, latent_dim=2).to(device)
+optimizer = Adam(model1.parameters(), lr=0.0001)         # fresh optimizer
+epochs = 1000
+beta = 1
+losses1 = train_epochs(epochs, model1, train_loader, optimizer, loss, beta = beta)
+
+
+# %% [markdown]
+# Let's have a look at reconstructions for model1: 
+
+# %%
+view_test_sample(model1, test_loader)
+
+
+# %% [markdown]
+# Let's compare the losses: 
+
+# %%
+def plot_losses_compare(loss_dicts, labels, colors=None):
+    """
+    loss_dicts: list of dicts, e.g. [losses_1, losses_2]
+    labels:     list of strings, e.g. ["Beta 0.001", "Beta 1"]
+    colors:     optional list of colors, e.g. ["blue", "red"]
+    """
+    if colors is None:
+        colors = plt.rcParams['axes.prop_cycle'].by_key()['color']  # default matplotlib colors
+
+    keys = list(loss_dicts[0].keys())
+    fig, axes = plt.subplots(1, len(keys), figsize=(6 * len(keys), 3))
+
+    for i, key in enumerate(keys):
+        for loss_dict, label, color in zip(loss_dicts, labels, colors):
+            axes[i].plot(loss_dict[key], label=label, color=color, marker='o', markersize=2)
+        axes[i].set_title(key)
+        axes[i].set_xlabel("Epoch")
+        axes[i].set_ylabel("Loss")
+        axes[i].grid(True, linestyle='--', alpha=0.6)
+        axes[i].legend()
+
+    plt.tight_layout()
+    plt.show()
+
+
+plot_losses_compare(
+    loss_dicts=[losses0, losses1],
+    labels=["Beta 0", "Beta high"],
+)
+
+
+# %% [markdown]
+# <div class="alert alert-block alert-success"><h2>Checkpoint</h2>
+#
+#
+# </div>
+
+# %% [markdown]
+# ### Part A.2.7: Apply the models to test images
+# #### Part A.2.7.1: Get latent space properties
+
+# %% [markdown]
+# Previously, we encoded one image into the latent space. Now we will encode the the entire test set. 
+# Below is a function that receives a model and a dataloader. 
+#
+# The function returns:
+# | Return value | Shape | Description |
+# |---|---|---|
+# | `mus` | `(10000, 2) (images, dims)` | latent mean for every test image |
+# | `logvars` | `(10000, 2) (images, dims)` | latent log-variance for every test image |
+# | `lbls` | `(10000,) (images,)` | digit label for every test image |
+# | `mu_mean` | `(10, 2) (digit class, dims)` | average latent position per digit class |
+
+# %%
+def get_latent_features(model, loader):
+    model.eval()
+    latents = []
+    logvars = []
+    labels = []
+
+    with torch.no_grad():
+        for x, lbl in loader:
+            x = x.to(device)
+            mu, logvar = model.encode(x)
+            latents.append(mu.cpu())
+            logvars.append(logvar.cpu())
+            labels.append(lbl)
+
+    mus = torch.cat(latents, dim=0) 
+    logvars = torch.cat(logvars, dim=0)
+    lbls = torch.cat(labels, dim=0)
+
+    mu_mean = torch.stack([mus[lbls == i].mean(dim=0) for i in range(10)])
+    return mus, logvars, lbls, mu_mean
+
+
+
+# %% [markdown]
+# Let's get these features for `model0` and `model1`: 
+
+# %%
+
+#get all latent features
+# model 0 
+mus_model0, logvars0, lbls0, mu_mean0 = get_latent_features(model0, tqdm(test_loader))
+
+# model 1 
+mus_model1, logvars1, lbls1, mu_mean1 = get_latent_features(model1, tqdm(test_loader))
+
+
+print(f"mu shape: {mus_model0.shape}, labels shape: {lbls0.shape}")
+
+# %% [markdown]
+# #### Part A.2.7.2: Visualize the latent spaces. 
+# **Let's plot the two dimensions of the latent space**.  
+# Every data-point represents one test-image. We plot the two latent dimensions (mu₁, mu₂) for `model0` and `model1` 
+
+# %%
+import numpy as np 
+
+def scatter_digits(ax, mus, lbls, mu_mean=None, alpha=1, CMAP = "tab10"):
+    """Colour-coded scatter, one series per digit so legend works."""
+
+    CMAP = plt.get_cmap(CMAP)
+
+    for d in range(10):
+        mask = lbls == d
+        ax.scatter(mus[mask, 0], mus[mask, 1], s=1, color=CMAP(d),
+                   label=str(d), rasterized=True, alpha=alpha)
+
+    if mu_mean is not None:
+        for d in range(10):
+            ax.scatter(*mu_mean[d], s=220, color="white", edgecolors="white", linewidths=3, marker="X", zorder=9)
+            ax.scatter(*mu_mean[d], s=80,  color=CMAP(d), edgecolors="black", linewidths=0.5, marker="X", zorder=10)
+            ax.annotate(str(d), xy=mu_mean[d], fontsize=8, fontweight="bold",
+                        ha="center", va="bottom", xytext=(0, 5),
+                        textcoords="offset points", zorder=11,
+                        bbox=dict(boxstyle="round,pad=0.15", fc="white", ec="none", alpha=0.8))
+
+    ax.legend(title="Digit", markerscale=6, ncol=2, fontsize=7, loc="best")
+    ax.set_aspect("equal")
+
+
+def scatter_with_normal(ax, mus, lbls, rnd_normal, mean, std):
+    """Digits on top of magenta N(0,1) cloud."""
+    ax.scatter(rnd_normal[:, 0], rnd_normal[:, 1],
+               s=1, alpha=0.25, color="magenta", label=f"N({mean},{std})", rasterized=True, zorder=5)
+    ax.scatter(mus[:, 0], mus[:, 1],
+               s=1, color="gray", label="mu", rasterized=True, zorder=2)
+    ax.legend(markerscale=6, ncol=2, fontsize=7, loc="best")
+    ax.set_aspect("equal")
+
+
+def plot_latent_digits(mus_model0, lbls0, mu_mean0, mus_model1, lbls1, mu_mean1):
+    """Latent space coloured by digit, with per-class centroids."""
+    fig, axes = plt.subplots(1, 2, figsize=(13, 5))
+    for ax, mus, lbls, mu_mean, title in [
+        (axes[0], mus_model0, lbls0, mu_mean0, "model0: β=0 latent space"),
+        (axes[1], mus_model1, lbls1, mu_mean1, "model1: β>0 latent space"),
+    ]:
+        scatter_digits(ax, mus, lbls, mu_mean=mu_mean)
+        ax.set_title(title, fontsize=11)
+        ax.set_xlabel("mu₁"); ax.set_ylabel("mu₂")
+    fig.suptitle("Latent space — coloured by digit", fontsize=13)
+    plt.tight_layout(); plt.show()
+
+
+
+
+def plot_latent_vs_normal(mus_model0, lbls0, mus_model1, lbls1, rnd_normal,
+                          clf0=None, clf1=None, mean=0, std=1, resolution=500):
+    fig, axes = plt.subplots(1, 2, figsize=(13, 5))
+
+    for ax, mus, lbls, clf, title in [
+        (axes[0], mus_model0, lbls0, clf0, f"β=0 vs N({mean},{std})"),
+        (axes[1], mus_model1, lbls1, clf1, f"β>>0 vs N({mean},{std})"),
+    ]:
+        # Decision boundaries (if a classifier is provided )
+        if clf is not None:
+            x_min, x_max = mus[:, 0].min() - 0.5, mus[:, 0].max() + 0.5
+            y_min, y_max = mus[:, 1].min() - 0.5, mus[:, 1].max() + 0.5
+            xx, yy = np.meshgrid(np.linspace(x_min, x_max, resolution),
+                                 np.linspace(y_min, y_max, resolution))
+            grid_preds = clf.predict(np.c_[xx.ravel(), yy.ravel()]).reshape(xx.shape)
+            # ax.contourf(xx, yy, grid_preds, levels=np.arange(-0.5, 10, 1),
+            #             cmap="tab10", alpha=0.25, zorder=0)
+            ax.contour(xx, yy, grid_preds, levels=np.arange(-0.5, 10, 1),
+                       colors="k", linewidths=0.4, zorder=1)
+
+        # Normal prior cloud + latent points
+        scatter_with_normal(ax, mus, lbls, rnd_normal, mean, std)
+        ax.set_title(title, fontsize=11)
+        ax.set_xlabel("mu₁"); ax.set_ylabel("mu₂")
+
+    fig.suptitle(f"Latent space vs N({mean},{std}) prior", fontsize=13)
+    plt.tight_layout()
+    plt.show()
+
+
+
+
+#call
+plot_latent_digits(mus_model0, lbls0, mu_mean0, mus_model1, lbls1, mu_mean1)
+
+mean = 0
+std = 1
+plot_latent_vs_normal(mus_model0, lbls0, mus_model1, lbls1, rnd_normal=np.random.normal(mean, std, size=(10000, 2)), mean = mean, std = std)
+
+
+# %% [markdown]
+# **top row**  
+# Each embedded image is color-coded by digit class. The marker x shows the per-class centroid – the average mu across all images of that digit. 
+#
+# **bottom row**  
+# Encoded digits (gray) are overlayed with 10000 samples from a standard normal distribution N(0,1) prior to assess how well the learned latent distribution matches it.
+
+# %% [markdown]
+# <div class="alert alert-block alert-warning"><h4> Questions </h4>
+# <ul>
+# <li>Look at the per-class centroids for a model of your choice. Do you think there's a reason why some centroids are closer to each other, while others are more distant? </li>
+# <li>Which model's latent space is more similar to a standard normal distribution? Why is this the case?</li>
+# </ul>
+# </div>
+
+# %% [markdown]
+# <div class="alert alert-block alert-success"><h2>Checkpoint</h2>
+#
+#
+# </div>
+
+# %% [markdown]
+# **visualizing `logvar`**  
+#
+# We plotted `mu`, but what about `logvar`? 
+#
+# So far we have only visualized `mu`, the center of each images distribution in latent space. 
+# The encoder also outputs `logvar`, which controls the spread of each images' distribution. 
+#
+# Below we will sample 100 values `z` from `mu` and `logvar`. 
+#
+# So each of the 10000 test-image is represented by 100 points – 100000 points in total. 
+
+# %%
+def sample_from_latents(mus, logvars, n_samples=100):
+    """
+    For each (mu, logvar) pair, draw n_samples via the reparameterization trick.
+    Returns z_samples of shape (N * n_samples, 2).
+    """
+    N, latent_dim = mus.shape
+    # expand to (N, n_samples, latent_dim)
+    mu_exp     = mus.unsqueeze(1).expand(-1, n_samples, -1)
+    logvar_exp = logvars.unsqueeze(1).expand(-1, n_samples, -1)
+
+    std = torch.exp(logvar_exp / 2)
+    eps = torch.randn_like(std)
+    z   = mu_exp + eps * std                        # (N, n_samples, 2)
+    return z.reshape(N * n_samples, latent_dim)     # (N*n_samples, 2)
+
+
+# ── sample ─────────────────────────────────────────────────────────────────
+n_samples = 100
+
+with torch.no_grad():
+    z_samples0 = sample_from_latents(mus_model0, logvars0, n_samples)
+    z_samples1 = sample_from_latents(mus_model1, logvars1, n_samples)
+
+# repeat each label n_samples times to match the expanded z array
+lbls0_rep = lbls0.repeat_interleave(n_samples)
+lbls1_rep = lbls1.repeat_interleave(n_samples)
+
+# ── plot ────────────────────────────────────────────────────────────────────
+fig, axes = plt.subplots(1, 2, figsize=(13, 7))
+
+ax = axes[0]
+scatter_digits(ax, z_samples0.numpy(), lbls0_rep, mu_mean=mu_mean0, alpha = 0.1)
+ax.set_title("β=0 — 100 samples per posterior", fontsize=11)
+ax.set_xlabel("z₁"); ax.set_ylabel("z₂")
+
+ax = axes[1]
+scatter_digits(ax, z_samples1.numpy(), lbls1_rep, mu_mean=mu_mean1, alpha = 0.1)
+ax.set_title("β=high — 100 samples per posterior", fontsize=11)
+ax.set_xlabel("z₁"); ax.set_ylabel("z₂")
+
+plt.suptitle("Sampled z", fontsize=13)
+plt.tight_layout()
+plt.show()
+
+# %% [markdown]
+# <div class="alert alert-block alert-warning"><h4> Questions </h4>
+# <ul>
+# <li>Which latent space looks more continuous? </li>
+# <li>Why is it important that the latent space follows a known distribution?</li>
+# </ul>
+# </div>
+
+# %% [markdown]
+# <div class="alert alert-block alert-success"><h2>Checkpoint</h2>
+#
+#
+# </div>
+
+# %% [markdown]
+# #### Part A.2.7.3: Classification in the latent space
+# The previous plots show clusters of numbers emerging. But can we quantify how well these clusters are separated? 
+#
+# Next, we train a **logistic regression classifier** on the latent space. 
+# This is asking how well a **simple linear model** can distinguish digits 
+# using only their latent coordinates (μ₁, μ₂) as features.
+#
+# If the clusters are truly well-separated, logistic regression — which can 
+# only draw straight decision boundaries — should achieve high accuracy.
+#
+# Note we're compressing the entire MNIST image (784 pixels) down to just 
+# **2 numbers**, so the accuracy also reflects how much class-relevant 
+# information is preserved in the bottleneck.
+
+# %%
+from sklearn.linear_model import LogisticRegression
+from sklearn.metrics import accuracy_score, confusion_matrix
+from sklearn.model_selection import train_test_split
+
+def logreg(mus, lbls, test_size=0.2, random_state=42):
+
+    mus_train, mus_val, lbls_train, lbls_val = train_test_split(
+        mus.numpy(), lbls.numpy(),
+        test_size=test_size,
+        random_state=random_state,
+        stratify=lbls.numpy()  # ensures each split has the same class proportions
+    )
+
+    # Fit only on training data
+    clf = LogisticRegression(max_iter=1000)
+    clf.fit(mus_train, lbls_train)
+
+    # Evaluate only on validation data
+    preds_val = clf.predict(mus_val)
+    accuracy = accuracy_score(lbls_val, preds_val)
+
+    unique_lbls = sorted(set(lbls_val))
+    conf_m = confusion_matrix(lbls_val, preds_val, labels=unique_lbls)
+
+    return accuracy, unique_lbls, conf_m, clf
+
+accuracy0, unique_lbls0, conf_m0, clf0 = logreg(mus_model0, lbls0)
+accuracy1, unique_lbls1, conf_m1, clf1 = logreg(mus_model1, lbls1)
+
+
+# %% [markdown]
+# We can visualize the decision boundaries: 
+
+# %%
+def plot_decision_boundaries(ax, clf, mus, lbls, title, resolution=500, CMAP="tab10"):
+    """Shade decision regions of a fitted classifier over the 2-D latent space."""
+    CMAP_obj = plt.get_cmap(CMAP)
+
+    x_min, x_max = mus[:, 0].min() - 0.5, mus[:, 0].max() + 0.5
+    y_min, y_max = mus[:, 1].min() - 0.5, mus[:, 1].max() + 0.5
+
+    xx, yy = np.meshgrid(
+        np.linspace(x_min, x_max, resolution),
+        np.linspace(y_min, y_max, resolution)
+    )
+
+    grid_preds = clf.predict(np.c_[xx.ravel(), yy.ravel()]).reshape(xx.shape)
+
+    # Shade decision regions (low alpha so data points stay visible)
+    ax.contourf(xx, yy, grid_preds, levels=np.arange(-0.5, 10, 1),
+                cmap=CMAP, alpha=0.25, zorder=0)
+
+    # Draw crisp decision boundaries
+    ax.contour(xx, yy, grid_preds, levels=np.arange(-0.5, 10, 1),
+               colors="k", linewidths=0.4, zorder=1)
+
+    # Overlay data points
+    scatter_digits(ax, mus, lbls)
+    ax.set_title(title, fontsize=11)
+    ax.set_xlabel("mu₁"); ax.set_ylabel("mu₂")
+    ax.set_xlim(x_min, x_max); ax.set_ylim(y_min, y_max)
+
+
+fig, axes = plt.subplots(1, 2, figsize=(13, 5))
+plot_decision_boundaries(axes[0], clf0, mus_model0.numpy(), lbls0.numpy(), "β=0 — decision boundaries")
+plot_decision_boundaries(axes[1], clf1, mus_model1.numpy(), lbls1.numpy(), "β>0 — decision boundaries")
+fig.suptitle("Logistic-regression decision boundaries in latent space", fontsize=13)
+plt.tight_layout(); plt.show()
+
+
+# %% [markdown]
+# Now let's plot the confusion matrix:
+
+# %%
+def confmatrix(ax, conf_m, accuracy, unique_lbls, title=""):
+    ax.imshow(conf_m)
+    ax.set_title(f"{title}\nAccuracy: {accuracy:.4f}")
+    ax.set_ylabel("True")
+    ax.set_xlabel("Predicted")
+    for i in range(conf_m.shape[0]):
+        for j in range(conf_m.shape[1]):
+            ax.text(j, i, conf_m[i, j],
+                    ha="center", va="center", color="w")
+    ax.set_xticks(unique_lbls)
+    ax.set_yticks(unique_lbls)
+
+
+fig, axes = plt.subplots(1, 2, figsize=(12, 5))
+
+confmatrix(axes[0], conf_m0, accuracy0, unique_lbls0, title="β=0")
+confmatrix(axes[1], conf_m1, accuracy1, unique_lbls1, title="β>>0")
+
+plt.suptitle("Logistic Regression on Latent Space (μ₁, μ₂)", fontsize=13)
+plt.tight_layout()
+plt.show()
+
+# %% [markdown]
+# <div class="alert alert-block alert-warning"><h4> Questions </h4>
+# <ul>
+# * Question: is one classification better than the other?  <br>
+# * If so, why do you think is the classification accuracy higher in one model than in the other?  
+# </ul>
+# </div>
+
+# %% [markdown]
+# <div class="alert alert-block alert-success"><h2>Checkpoint</h2>
+#
+#
+# </div>
+
+# %% [markdown]
+# ### A.2.8. Sample from the latent space
+# So far, we have looked at the latent space. Next, we look at what the decoder produces given positions in the latent space 
+
+# %% [markdown]
+# Function `gen_mean_numbers` takes `mu_mean`, the average latent position of all test images of digit i. 
+# These correspond to the centroids shown in the plots above. 
+#
+# The result is the most typical representation of each digit as embedded by the model. 
+#
+# In a latent-space with well-separated clusters, each centroid should look like a clean version of the digit.  
+
+# %%
+# Uncomment and run if you want to see the latent-space again: 
+plot_latent_digits(mus_model0, lbls0, mu_mean0, mus_model1, lbls1, mu_mean1)
+
+
+# %%
+# Plot average latent vectors 
+
+def gen_mean_numbers(model, mu_mean, title):
+    fig, ax = plt.subplots(1, 10, figsize = (20, 4))
+    with torch.no_grad():
+        for i in range(10):
+            gen = model.decode(mu_mean[i].to(device))
+
+            ax[i].imshow(gen.detach().cpu().squeeze().numpy(), cmap = "Grays")
+            ax[i].set_title(f"Mean for {i}")
+    fig.suptitle(title)
+    plt.tight_layout()
+
+
+# gen_mean_numbers(model0, mu_mean0, title="Model 0") # model with beta = 0
+gen_mean_numbers(model1, mu_mean1, title="Model 1") # model with beta >> 0
+
+# %% [markdown]
+# #### Interpolate between digits in the latent space 
+# Above, we have seen "mean" representations for each digit, corresponding to centroids of each digit class in the latent space.  
+# We can decode points sampled along the straight line between two centroids in latent space. A visualization will be plotted below.  
+#
+
+# %% [markdown]
+# <div class="alert alert-block alert-info"><h2>Task</h2>
+# Run the code below without modifications. You should see reconstructions of interpolated images along the path between the centroids of 0 and 6. <br> 
+# Next, choose your own two digit centroids. Latent space permitting, try three types of paths:
+#
+# * Gap — the line passes through an empty region of the latent space   <br>  
+# * Traversal — the line passes through one or more other digit clusters<br>  
+# * Clean — the line connects two digits directly, with no gaps and no other clusters in between<br>  
+#
+#
+# For each case: can you explain what the decoded images look like, and why?
+# </div>
+
+# %%
+# TODO: pick digits 
+digit_a = 0
+digit_b = 6
+steps = 12 # number of times to sample 
+
+
+def plot_latent_digits_interp(mus_model0, lbls0, mu_mean0, mus_model1, lbls1, mu_mean1, digit_a, digit_b):
+    """Latent space coloured by digit, with a line connecting two class centroids."""
+    fig, axes = plt.subplots(1, 2, figsize=(10, 5))
+    for ax, mus, lbls, mu_mean, title in [
+        (axes[0], mus_model0, lbls0, mu_mean0, "β=0 latent space"),
+        (axes[1], mus_model1, lbls1, mu_mean1, "β=high latent space"),]:
+        scatter_digits(ax, mus, lbls, mu_mean=mu_mean)
+
+        # ── line between the two class centroids ──────────────────────────
+        pts = np.stack([mu_mean[digit_a], mu_mean[digit_b]])
+        ax.plot(pts[:, 0], pts[:, 1], c="black", lw=2, ls="--", zorder=20)
+
+        ax.set_title(title, fontsize=11)
+        ax.set_xlabel("mu₁"); ax.set_ylabel("mu₂")
+
+    fig.suptitle(f"Latent space — interpolation path {digit_a} → {digit_b}", fontsize=13)
+    plt.tight_layout(); plt.show()
+
+
+
+
+# interpolate between two digits 
+def interpolate(digit_a, digit_b, mu_mean, model, title, steps = 10):
+
+    mu_a = mu_mean[digit_a]
+    mu_b = mu_mean[digit_b]
+
+    _, ax = plt.subplots(1, steps, figsize = (20, 3))
+
+    with torch.no_grad():
+        for i, weight_b in enumerate(np.linspace(0, 1, steps)):
+            weight_a = 1 - weight_b
+            trajectory = weight_a * mu_a + weight_b * mu_b
+            trajectory = trajectory.to(device)
+            gen = model.decode(trajectory)
+
+            ax[i].imshow(gen.detach().cpu().squeeze().numpy(), cmap = "Grays")
+            ax[i].set_title(f"{weight_a:.1f} x $\mu_{digit_a}$, {weight_b:.1f} x $\mu_{digit_b}$")
+    plt.suptitle(title)
+    plt.tight_layout()
+
+
+
+plot_latent_digits_interp(mus_model0, lbls0, mu_mean0, mus_model1, lbls1, mu_mean1, digit_a=digit_a, digit_b=digit_b)
+
+interpolate(digit_a, digit_b, mu_mean0, model0, "Beta = 0", steps = steps)
+interpolate(digit_a, digit_b, mu_mean1, model1, "Beta > 0", steps = steps)
+
+# %% [markdown]
+# <div class="alert alert-block alert-info"><h2>Bonus task</h2>
+# Change coordinates <code>z = [mu₁, mu₂]</code> to decode any point in the latent spaces.  <br>
+# <code>z</code> will appear as a magenta star.  
+# </div>
+
+# %%
+# TODO: pick a point in the latent space
+MU1 = -20
+MU2 = -12
+z = [MU1, MU2] 
+
+def decode_point(z, model0, model1, mus_model0, lbls0, mu_mean0, mus_model1, lbls1, mu_mean1):
+    z_tensor = torch.tensor(z, dtype=torch.float32).unsqueeze(0).to(device)
+
+    fig, axes = plt.subplots(1, 4, figsize=(12, 8))
+
+    for ax_scatter, ax_img, model, mus, lbls, mu_mean, title in [
+        (axes[0], axes[1], model0, mus_model0, lbls0, mu_mean0, "β=0"),
+        (axes[2], axes[3], model1, mus_model1, lbls1, mu_mean1, "β>0"),
+    ]:
+        # latent space with marked point
+        scatter_digits(ax_scatter, mus, lbls, mu_mean=mu_mean)
+        ax_scatter.scatter(*z, s=200, color="magenta", marker="*", zorder=20)
+        ax_scatter.set_title(f"{title} latent space")
+        ax_scatter.set_xlabel("mu₁"); ax_scatter.set_ylabel("mu₂")
+
+        # decoded image
+        with torch.no_grad():
+            gen = model.decode(z_tensor)
+        ax_img.imshow(gen.cpu().squeeze(), cmap="gray")
+        ax_img.set_title(f"{title} decoded")
+        ax_img.axis("off")
+
+    plt.tight_layout()
+    plt.show()
+
+decode_point(z, model0, model1, mus_model0, lbls0, mu_mean0, mus_model1, lbls1, mu_mean1)
+
+# %% [markdown]
+# <div class="alert alert-block alert-success"><h2>Checkpoint</h2>
+#
+#
+# </div>
+
+# %% [markdown]
+# ## A.2.9: Higher-dimensional latent-spaces
+#
+# So far, we've only looked at a models with two latent dimensions. 
+# It is easier to visualize and intuitively understand two dimensions, but two numbers may not be enough to capture all meaningful variation of the dataset. 
+# Digits may overlap in latent-space, because two dimensions do not offer enough room to separate them out. 
+#
+# Let's train a model with higher-dimensional latent sapce. Below, we
+# * Instantiate a new variational autoencoder model and name it `model2`
+# * Instantiate a new optimizer 
+# * Pass `beta = 1`
+# * Train your new model for `epochs = 1000`
+
+# %% [markdown]
+# <div class="alert alert-block alert-info"><h2>Task</h2>
+#
+# Choose `latent_dim = n`, where n is bigger than 5 and smaller than 100
+#
+# </div>
+#
+
+# %%
+latent_dim = ...
+
+model2 = VariationalAutoEncoder(w, h, latent_dim = latent_dim).to(device)
+optimizer = Adam(model2.parameters(), lr = 0.0001)
+
+epochs = 1000
+beta = 1 
+
+train_epochs(epochs, model2, train_loader, optimizer, loss, beta = beta);
+
+# %% tags=["solution"]
+latent_dim = 99
+
+model2 = VariationalAutoEncoder(w, h, latent_dim = latent_dim).to(device)
+optimizer = Adam(model2.parameters(), lr = 0.0001)
+
+epochs = 1000
+beta = 1 
+
+train_epochs(epochs, model2, train_loader, optimizer, loss, beta = beta);
+
+# %% [markdown]
+# Let's have a look at a few reconstructions: 
+
+# %%
+view_test_sample(model2, test_loader)
+
+# %% [markdown]
+# Let's fast-forwarde – get latent features
+
+# %%
+mus2, logvars2, lbls2, mu_mean2 = get_latent_features(model2, tqdm(test_loader))
+
+print(mus2.shape) # N samples, latent dims
+print(logvars2.shape) # N samples, latent dims
+print(lbls2.shape) # N samples
+print(mu_mean2.shape) # N digits, latent dims
+
+
+# %% [markdown]
+# #### Dimensionality reduction on the latent space 
+# We now have too many latent dimensions to easily visualize. 
+#
+# We therefore apply **UMAP** (Uniform Manifold Approximation and Porjection). UMAP is a dimensionality reduction technique that non-linearly projects the high-dimensional latent space to 2D.  
+# It attempts to keep points that are close in high-dimensions close in their 2D projection. 
+#
+# `run_umap` performs this projection on `mus2` (the posterior mean in the latent space of all test images) and `mu_mean`, the per-class centroid in the latent space. 
+#
+# Note that UMAP is a visualization tool. We can use it to build intuition, but should not draw conclusions about the geometry of the latent space. 
+
+# %%
+from umap import UMAP
+from matplotlib.colors import ListedColormap
+
+def run_umap(latents, n_components=2, random_state=42, n_neighbors=15, min_dist=0.1, means=None):
+    reducer = UMAP(n_components=n_components, random_state=random_state,
+                   n_neighbors=n_neighbors, min_dist=min_dist)
+    if means is not None:
+        combined = np.concatenate([latents.numpy(), means.numpy()])
+        combined_2d = reducer.fit_transform(combined)
+
+        # Split back into latents and means
+        mu_2d = combined_2d[:len(latents)]        # (10000, 2)
+        mu_mean_2d = combined_2d[len(latents):]  # (10, 2)
+        return mu_2d, mu_mean_2d
+
+    mu_2d = reducer.fit_transform(latents.numpy())
+    return mu_2d
+
+
+def plot_umap(mu_2d, labels,
+              cmap='tab10', alpha=0.6, s=10, centers=None, center_labels=None):
+
+    ticks = np.unique(labels)
+    base_cmap = plt.get_cmap(cmap)
+    colors_n = base_cmap(np.linspace(0, 1, np.max(ticks) + 1))
+    new_cmap = ListedColormap(colors_n)
+
+    plt.figure(figsize=(10, 8))
+    scatter = plt.scatter(mu_2d[:, 0], mu_2d[:, 1], c=labels, cmap=new_cmap, alpha=alpha, s=s)
+    plt.colorbar(scatter, ticks=ticks)
+
+    if centers is not None:
+        for i, label in enumerate(center_labels):
+            plt.text(centers[i, 0], centers[i, 1], s=str(label), backgroundcolor="white", size='large')
+        plt.scatter(centers[:, 0], centers[:, 1], s=50, marker="X", c="k", zorder=10000)
+    plt.axis('off')
+
+
+# %%
+mu_2d_2, mu_means_2d_2 = run_umap(mus2, means = mu_mean2)
+plot_umap(mu_2d_2, lbls2, cmap = "tab10", centers = mu_means_2d_2, center_labels=range(10))
+
+# %% [markdown]
+# Clusters look well-separated, but we should verify with logistic regression.  
+# Below, we call the logistic regression function again and plot the confusion matrix.  
+
+# %%
+accuracy2, unique_lbls2, conf_m2, clf2 = logreg(mus2, lbls2)
+
+
+fig, axes = plt.subplots(1, 1, figsize=(5, 5))
+
+confmatrix(axes, conf_m2, accuracy2, unique_lbls2, title= f"β=1, latent_dims = {latent_dim}")
+
+plt.suptitle("Logistic Regression on Latent Space ", fontsize=13)
+plt.tight_layout()
+plt.show()
+
+# %% [markdown]
+# <div class="alert alert-block alert-success"><h2>Checkpoint</h2>
+#
+#
+# </div>
+
+# %% [markdown]
+# # PART B: Explainable AI (XAI)
+# ## Part B.1: Setup
 #
 # In this part of the notebook, we will load the same dataset as in the previous exercise.
 # We will also learn to load one of our trained classifiers from a checkpoint.
@@ -345,7 +1967,7 @@ for attr, im, lbl in zip(attributions_blurred, x.cpu().numpy(), y.cpu().numpy())
     visualize_color_attribution(attr, im)
 
 # %% [markdown] tags=[]
-# <div class="altert alert-block alert-warning"><h4> Questions </h4>
+# <div class="alert alert-block alert-warning"><h4> Questions </h4>
 # <ul>
 # <li>What baseline do you like best so far? Why?</li>
 # <li>Why do you think some baselines work better than others?</li>
